commander_client/requirements.txt
# GridPass Commander Client
# A lightweight PC service for iRacing telemetry and control

# HTTP client
requests>=2.28.0

# Environment configuration
python-dotenv>=1.0.0

# iRacing SDK
pyirsdk>=1.3.0

# GUI (PyQt6)
PyQt6>=6.4.0

# Build tools (optional - for creating executables)
# pyinstaller>=5.0

